requirements.txt
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
waitress==2.1.2
Werkzeug==3.0.1
numpy==1.19.1

